Adding test for sourceStringValidation
