+++ src/UI/api/utils/escapeRegex.ts
@@ -1,0 +1,11 @@
+export function escapeRegex(inputField: string) {
+  if (typeof inputField === 'boolean') {
+    return inputField;
+  } else {
+    let escapeinputField = inputField.replace(
+      /[-[\]/{}()*+?.,\\^$|#"]/g,
+      '\\$&'
+    );
+    return escapeinputField;
+  }
+}
