+++ src/UI/api/utils/sourceStringValidation.test.ts
@@ -1,22 +1,16 @@
-import { escapeRegex, sourceStringValidation } from './sourceStringValidation';
+import { sourceStringValidation } from './sourceStringValidation';
 
 describe('sourceStringValidation', () => {
-  it('adds escape characters where appropriate', () => {
-    const testStringNone: string = 'This is a test string';
-    // eslint-disable-next-line prettier/prettier
-    const testStringBackslashEmbedded: string = 'Thi\\s i\\s a t\\est str\\ing';
-    const testStringBackslashEdge: string = '\\This is\\ a \\test string\\';
-    const testStringDoubleQuotesEmbedded: string =
-      'Th"i"s is a t"e"st s"trin"g';
-    const testStringDoubleQuotesEdge: string = '"This" is "a" test "string"';
-    expect(escapeRegex(testStringNone)).toEqual('This is a test string');
-    // eslint-disable-next-line prettier/prettier
-    expect(escapeRegex(testStringBackslashEmbedded)).toEqual('Thi\\\\s i\\\\s a t\\\\est str\\\\ing');
-    // eslint-disable-next-line prettier/prettier
-    expect(escapeRegex(testStringBackslashEdge)).toEqual('\\\\This is\\\\ a \\\\test string\\\\');
-    // eslint-disable-next-line prettier/prettier
-    expect(escapeRegex(testStringDoubleQuotesEmbedded)).toEqual('Th\\\"i\\\"s is a t\\\"e\\\"st s\\\"trin\\\"g');
-    // eslint-disable-next-line prettier/prettier
-    expect(escapeRegex(testStringDoubleQuotesEdge)).toEqual('\\\"This\\\" is \\\"a\\\" test \\\"string\\\"');
+  const testSourceStringObject: any = {
+    test1: 'This is a test string',
+    test2: 'Thi\\s i\\s a t\\est str\\ing',
+    test3: 'Thi\\s i\\s a t\\est str\\ing',
+  };
+  it('call on escapeRegex when required to process string', () => {
+    jest.mock('./escapeRegex', () => ({
+      escapeRegex: jest.fn(),
+    }));
+    sourceStringValidation(testSourceStringObject);
+    expect(escapeRegex).toBeCalledTimes(2);
   });
 });
